Credit notes (via Stripe credit notes) (#112)
Switch credit note creation from refunds to actual credit notes

Refunds are not a reliable way of creating credit notes, as they do
not take discounts or VAT into account. Now that we are on an API
version that supports the credit notes feature of Stripe, we can
switch to the `credit_note.created` webhook instead.

* Refactor view model for credit notes
* Update accounting export scripts
* Add backfill credit notes data migration script

Remember to run the backfill script on deploy if you were using Billbo before this release.