Ensure credit note metadata matches up with original invoice
Between the time that an invoice was made and the time when the
creditnote is actually issued, the VAT related metadata could have
been changed or updated. Stripe does the right thing with regards
to the actual VAT amounts, but since we're using metadata in Billbo
we need to manage that data ourselves.

VAT-related metadata is now taken from the original invoice to
ensure consistency.

Closes #116